feat: safety gate system + model improvements across nexora-ai
- Add GlobalSafetyGate with mandatory SPARO alignment check
- Add SandboxConfig for Cipher and Genesis isolation
- Add AuditTrail for Nexum multi-agent decisions
- Add ConsentToken for Aether psychological profiling
- Add CapabilityLock to prevent runtime downgrade of Security/Alignment
- Various model architecture refinements and bug fixes

+++ crates/foundation/src/models/aether/capabilities.rs
@@ -10,6 +10,7 @@ use crate::shared::{
 };
 
 /// NXR-ÆTHER Capabilities Manager
+#[derive(Clone)]
 pub struct AetherCapabilities {
     /// Capability vector
     vector: CapabilityVector,
@@ -479,7 +480,7 @@ impl AetherCapabilities {
             if let Some(capability) = new_capabilities.vector.capabilities.get_mut(domain) {
                 // Simulate improvement by increasing score
                 let current_score = capability.score();
-                let new_score = (current_score + improvement).min(1.0);
+                let new_score = (current_score + improvement).min(1.0_f32);
                 
                 // Update capability metrics
                 capability.metrics.insert("simulated_score".to_string(), new_score);
@@ -497,7 +498,7 @@ impl AetherCapabilities {
             psychological_analysis: self.performance_metrics.psychological_analysis_accuracy,
             cultural_adaptation: self.performance_metrics.cultural_adaptation_accuracy,
             support_generation: self.performance_metrics.support_generation_quality,
-            response_efficiency: (1000.0 / self.performance_metrics.avg_response_time_ms).min(1.0),
+            response_efficiency: ((1000.0 / self.performance_metrics.avg_response_time_ms).min(1.0)) as f32,
         }
     }
